mysite/requirements.txt
aiohttp==3.8.6
aioredis==2.0.1
aiosignal==1.3.1
annotated-types==0.6.0
arrow==1.3.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
binaryornot==0.4.4
certifi==2023.7.22
cffi==1.16.0
channels==4.0.0
channels-redis==4.1.0
chardet==5.2.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
constantly==15.1.0
cookiecutter==2.4.0
cryptography==41.0.4
daphne==4.0.0
Django==4.2.6
django-environ==0.11.2
django-tailwind==3.6.0
frozenlist==1.4.0
gunicorn==21.2.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
influxdb-client==1.38.0
Jinja2==3.1.2
mariadb==1.1.7
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
msgpack==1.0.7
multidict==6.0.4
mysql==0.0.3
mysqlclient==2.2.0
packaging==23.2
paho-mqtt==1.6.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.13
pydantic_core==2.10.1
Pygments==2.16.1
pymodbus==3.5.4
pyOpenSSL==23.2.0
python-dateutil==2.8.2
python-decouple==3.8
python-slugify==8.0.1
pytz==2023.3.post1
PyYAML==6.0.1
reactivex==4.0.4
redis==5.0.1
requests==2.31.0
rich==13.6.0
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
tailwind==3.1.5b0
text-unidecode==1.3
Twisted==23.8.0
txaio==23.1.1
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
yarl==1.9.2
zope.interface==6.1
